Add TCP flow support to the LP Solver
The LP Solver was updated to add support for TCP and  the ACK flows.
The XML file output by the LP Solver was updated to include the source port, ONLY for TCP flows. Udp flows have remain unchanged.

The LGF was updated to add a TCP Flow Id value at the end of the flow ONLY if it is an ACK flow. For all the other flow types this value will be ignored.

A check was set in place to make sure that the ACK flow refers to a valid TCP flow and that its source and destination are the opposite of the TCP data flow.