Fix bug when the includeAllKEqualCostPaths option is set
By mistake, the paths with cost larger than the kth path were being kept. The problem was due to the return variables being set in reverse to what they were required to be. A return value of true means that the element will be removed, the program's logic was that true means keep the item.