Fix indefinite loop in KSP when all paths of a flow were found
The KSP algorithm would loop indefinitely until an overflow error would cause the program to crash when all the paths of a flow have been discovered. A fix is implemented where if the number of paths found does not change from the current run and the next, the program will stop searching for more paths since all paths would have been discovered.