Update the unconstrained LpSolver to follow the new LGF file format
The unconstrained LpSolver is updated to work using the new and improved LGF file. This means that the LpSolver does not have any knowledge of ACK flows. This is to bring it on par with the rest of the solvers. If we would need to run ns3 simulations on these solutions we would required the use of the KSP file to get the AckPath.

Note the Flow::dataRate attribute is renamed to Flow::allocatedDataRate because it represents exactly what the variable is storing.