Fix issue in GA Ns3 input file generator + Improve log messages
The GA Ns3 input file generator had an issue in the new selection
mechanisms where the delay metric was not being set when a max flow
solution was found.
The output log messages were improved to include the index of the chosen
solution as well as the actual values of the objective metrics.