Fix indentation in the GA that was causing a crash
The log_info command needed to be run only if a pclp result file is
given. With the indentation setting as it was, it was being run all the
time. Since it refers to the chromosomeIndex variable that is created
only when a pclp result file is given the GA would crash. This has now
been fixed by moving the log_info line to the correct indentation group.